I've added some documentation to our project. Each package in MLH now has package-info.java which describes the purpose of the package. Also, I moved both  FileConstants and UIConstants to the Constants package.
