+++ MLH/src/android/mlh/aidl/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * I suppose this package consists of all the stuff
+ * that makes adding plug-ins possible?
+ * I don't know how it works, 
+ * so I'll let Ostrovsky handle this particular description.
+ * TODO: add description
+ */
+
+package android.mlh.aidl;
