+++ MLH/src/android/mlh/bl/files/FileManager.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 import android.content.Context;
 import android.mlh.bl.tasks.Task;
-import android.mlh.utils.constants.FileConstatns;
+import android.mlh.constants.FileConstatns;
 import android.util.Log;
 
 /** 
@@ -20,19 +20,35 @@ import android.util.Log;
  * is used for serialization.
  * PREFIX, EXT constants are defined in android.mlh.utils.constants.
  * Also deletes Serialized Tasks from the storage.
+ * We are using Singleton pattern.
+ * GetInstance static method receives Context, so the FileManager will always be 
+ * with the right context.
  */
 public class FileManager {
 	
+	/** FileManager instance - only one*/
 	private static FileManager _instance = null;
+	
+	/** context - so the FileManager will be updated constantly */
 	private Context mContext;
 	
 	private FileManager() {
 	}
 	
+	/**
+	 * Switching context of file manager 
+	 * @param context - Context
+	 */
 	private void switchContext(Context context) {
 		mContext = context;
 	}
 	
+	/**
+	 * Serializes the given Task, and writes it to a file.
+	 * Basically, saves the Task obj as a file.
+	 * @param task - Task
+	 * @throws IOException
+	 */
 	public void saveTask(Task task) throws IOException {
 		String filename = FileConstatns.TASK_FILE_NAME_PREFIX + task.getName() + 
 				FileConstatns.OBJECT_FILE_NAME_EXT;
@@ -48,6 +64,14 @@ public class FileManager {
         fileOut.close();
 	}
 	
+	/**
+	 * De-serializes Task from file to obj,
+	 * given the Task's name. 
+	 * @param taskName
+	 * @return the Task that was de-serialized
+	 * @throws IOException
+	 * @throws ClassNotFoundException
+	 */
 	public Task getTask(String taskName) throws IOException, ClassNotFoundException {
 		String filename = taskName + 
 				FileConstatns.OBJECT_FILE_NAME_EXT;
@@ -63,12 +87,25 @@ public class FileManager {
         return task;
 	}
 	
+	/**
+	 * Deletes the serialized Task - e.i. deletes file
+	 * that holds serialized Task.
+	 * @param taskName - the name of the Task
+	 * @throws IOException
+	 * @throws ClassNotFoundException
+	 */
 	public void deleteTask(String taskName) throws IOException, ClassNotFoundException {
 		File file = new File(getTaskPath(taskName));
         
         file.delete();
 	}
 	
+	/**
+	 * Returns path to the file which holds serialized Task
+	 * @param taskName
+	 * @return on success - path to the file which holds serialized Task,
+	 * or 'PATH NOT FOUND' on failure
+	 */
 	public String getTaskPath(String taskName) {
 		String filename = taskName + 
 				FileConstatns.OBJECT_FILE_NAME_EXT;
@@ -89,6 +126,10 @@ public class FileManager {
 		return "PATH NOT FOUND";
 	}
 	
+	/**
+	 * Returns current directory where serialized Tasks are stored
+	 * @return current directory where serialized Tasks are stored
+	 */
 	private File getTasksDir() {
 		return mContext.getFilesDir();
 	}
@@ -99,7 +140,7 @@ public class FileManager {
 	 * - from these files filter ones that match TASK_FILE_NAME_PREFIX%name%OBJECT_FILE_NAME_EXT
 	 * - each file is a serialized task (e.g. list of Experiments)
 	 * - returns list of filenames (i.e. tasks)
-	 * @return 
+	 * @return List of filenames in the current Task directory
 	 */
 	public List<String> getSavedTasks() {
 		ArrayList<String> savedTasks = new ArrayList<String>();
@@ -118,6 +159,12 @@ public class FileManager {
 		return savedTasks;
 	}
 	
+	/**
+	 * Returns reference to the FileManager.
+	 * Uses lazy instantiation.
+	 * @param context - given context
+	 * @return reference to the FileManager
+	 */
 	public static FileManager getInstance(Context context) {
 		if (_instance == null) {
 			_instance = new FileManager();
