+++ MLH/src/android/mlh/bl/files/package-info.java
@@ -1,0 +1,12 @@
+/**
+ * Contains classes for working with files.
+ * To be more specific: in context of this package
+ * a file is a serialized object. 
+ * So, basically this package provides classes
+ * that are associated with serialization/de-serialization
+ * of objects.
+ * The main class is FileManager, which exposes
+ * the serialization/de-serialization public methods.
+ */
+
+package android.mlh.bl.files;
