+++ MLH/src/android/mlh/bl/tasks/Task.java
@@ -9,6 +9,14 @@ import android.mlh.aidl.Experiment;
  * Basic task class. 
  * Stores list of experiments.
  * Instance is serialized into a file by FileManager.
+ * 
+ * This class is intended to store all the information for some task
+ * that user defined for himself.
+ * A task has a name, a type and a number of 'tries' which are called Experiments.
+ * User creates a task of some type, gives it a name and eventually fills it with 
+ * Experiments - descriptions of attempts to accomplish the Task. 
+ * Since there could be several experiments for the Task, it holds the Experiment
+ * currently selected by the user. 
  */
 public class Task implements Serializable{
 	
@@ -16,11 +24,19 @@ public class Task implements Serializable{
 	private String mName, mType, mPluginKey;
 	private List<Experiment> mExperiments;
 	
-	/** When we choose the task, we change it to the number (index in MainActivity.savedTasks arrayList)*/
+	/** Holds the index of currently selected Experiment (if any)
+	 *  This value is -1 by default/when no Experiment is selected.
+	 */
 	private int mCurrExperiment = -1;
 	
 	public static final int CURRENT_EXPERIMENT_NOT_DEFINED = -1;
 	
+	/**
+	 * Default constructor. Constructs a new Task.
+	 * @param name - task name
+	 * @param type - task type (what do you mean by 'type'?)
+	 * @param aPluginKey - what is that?
+	 */
 	public Task(String name, String type, String aPluginKey) {
 		setName(name);
 		setType(type);
@@ -29,14 +45,26 @@ public class Task implements Serializable{
 		mExperiments = new ArrayList<Experiment>();
 	}
 	
+	/**
+	 * Append the Experiment to the task.
+	 * @param experiment - given Experiment
+	 */
 	public void addExperiment(Experiment experiment) {
 		mExperiments.add(experiment);
 	}
 	
+	/**
+	 * Get list of all Experiments associated with this task
+	 * @return list with all Experiments
+	 */
 	public List<Experiment> getExperiments() {
 		return mExperiments;
 	}
 	
+	/**
+	 * Sets the type of the Task
+	 * @return
+	 */
 	public String getType() {
 		return mType;
 	}
@@ -45,26 +73,50 @@ public class Task implements Serializable{
 		this.mType = aType;
 	}
 	
+	/**
+	 * Returns the name of this Task
+	 * @return name of this Task
+	 */
 	public String getName() {
 		return mName;
 	}
-
+	
+	/**
+	 * Sets new name for this Task
+	 * @param aName - name for the Task
+	 */
 	public void setName(String aName) {
 		this.mName = aName;
 	}
 	
+	/**
+	 * 
+	 * @return
+	 */
 	public String getPluginKey() {
 		return mPluginKey;
 	}
-
+	
+	/**
+	 * TODO: I don't know what this [parameter does!
+	 * @param aPluginKey
+	 */
 	public void setPluginKey(String aPluginKey) {
 		this.mPluginKey = aPluginKey;
 	}
 	
+	/**
+	 * Get the Experiment which is currently selected by the user
+	 * @return Experiment that is selected - the current experiment
+	 */
 	public int getCurrentExperiment() {
 		return mCurrExperiment;
 	}
-
+	
+	/**
+	 * Sets given experiment to current (as currently selected by the user)
+	 * @param experiment - index of the Experiment in the Experiments list
+	 */
 	public void setCurrentExperiment(int experiment) {
 		this.mCurrExperiment = experiment;
 	}
