+++ MLH/src/android/mlh/bl/tasks/TaskManager.java
@@ -2,8 +2,14 @@ package android.mlh.bl.tasks;
 
 import android.mlh.aidl.Experiment;
 
+/** Holds reference to the Task that is currently selected by the user.
+ * If no Task is selected - reference is null.
+ * Holds only one task at the time.
+ * This class is a singleton with lazy instantiation.
+ */
 public class TaskManager {
 	
+	/** Single instance */
 	private static TaskManager _instance = null;
 	
 	private Task currTask;
@@ -11,6 +17,11 @@ public class TaskManager {
 	private TaskManager() {
 	}
 	
+	/**
+	 * Returns reference to the TaskManager instance.
+	 * Instantiates TaskManager in lazy manner.
+	 * @return reference to the TaskManager instance.
+	 */
 	public static TaskManager getInstance() {
 		if (_instance == null) {
 			_instance = new TaskManager();
@@ -19,10 +30,18 @@ public class TaskManager {
 		return _instance;
 	}
 	
+	/**
+	 * Saves given task as current - i.e. selected by the user.
+	 * @param task - given Task
+	 */
 	public void setCurrentTask(Task task) {
 		currTask = task;
 	}
 	
+	/**
+	 * Returns reference to currently selected Task
+	 * @return currently selected Task
+	 */
 	public Task getCurrentTask() {
 		return currTask;
 	}
