+++ MLH/src/android/mlh/bl/tasks/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * This package contains the Task class 
+ * and all classes that are associated with it.
+ */
+
+package android.mlh.bl.tasks;
+
+
