+++ MLH/src/android/mlh/constants/package-info.java
@@ -1,0 +1,16 @@
+/**
+ * This package contains contains classes with constants.
+ */
+/*
+ * 
+ * TODO: each class with constants has to be build like this:
+ *  - private constructor
+ *  - public static final fields
+ *  - in Target class we use static import: import static Constants.PLANCK_CONSTANT;
+ *  
+ *  That way we can use constants in Target class without
+ *  need to reference the namespace.
+ *
+ *	Maybe each constants class should be seated in the package of its usage?
+ */
+package android.mlh.constants;
