+++ MLH/src/android/mlh/ui/MainActivity.java
@@ -22,7 +22,7 @@ import android.mlh.bl.files.FileManager;
 import android.mlh.bl.plugins.PluginManager;
 import android.mlh.bl.tasks.Task;
 import android.mlh.bl.tasks.TaskManager;
-import android.mlh.ui.constants.UIConstatns;
+import android.mlh.constants.UIConstatns;
 import android.os.Bundle;
 import android.os.IBinder;
 import android.util.Log;
