+++ MLH/src/android/mlh/ui/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Here are all the UI-associated classes.
+ * All the Activity classes are here, including the main one
+ */
+
+package android.mlh.ui;
